+++ src/Subsystems/PneumaticsSubsytem.cpp
@@ -1,0 +1,21 @@
+/*----------------------------------------------------------------------------*/
+/* Copyright (c) 2017-2018 FIRST. All Rights Reserved.                        */
+/* Open Source Software - may be modified and shared by FRC teams. The code   */
+/* must be accompanied by the FIRST BSD license file in the root directory of */
+/* the project.                                                               */
+/*----------------------------------------------------------------------------*/
+
+#include "PneumaticsSubsystem.h"
+
+#include "../RobotMap.h"
+
+PneumaticsSubsystem::PneumaticsSubsystem()
+    : frc::Subsystem("PneumaticsSubsystem") {}
+
+void PneumaticsSubsystem::InitDefaultCommand() {
+	// Set the default command for a subsystem here.
+	// SetDefaultCommand(new MySpecialCommand());
+}
+
+// Put methods for controlling this subsystem
+// here. Call these from Commands.
